% Auto-generated: do not edit by hand
\name{''EditableGraph}

\alias{''EditableGraph}

\title{EditableGraph component}

\description{
It takes a for-loop to delete multiples traces, and then append the new ones.
}

\usage{
''EditableGraph(id=NULL, aim=NULL, data=NULL, style=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{aim}{Character. The ID used in the dcc.Graph to modify.}

\item{data}{Unnamed list. A list of dic with the data to update}

\item{style}{Named list. A object with the data style}
}
